PolygonLayer converted to typescript.
