+++ modules/layers/src/index.ts
@@ -40,3 +40,4 @@ export {default as _TextBackgroundLayer} from './text-layer/text-background-laye
 
 // Types
 export type {ScatterplotLayerProps} from './scatterplot-layer/scatterplot-layer';
+export type {PolygonLayerProps} from './polygon-layer/polygon-layer';
