+++ modules/layers/src/polygon-layer/polygon-layer.ts
@@ -18,16 +18,190 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import {CompositeLayer, createIterable, log} from '@deck.gl/core';
+import {
+  Accessor,
+  AccessorFunction,
+  Color,
+  CompositeLayer,
+  CompositeLayerProps,
+  createIterable,
+  log,
+  Unit
+} from '@deck.gl/core';
 import SolidPolygonLayer from '../solid-polygon-layer/solid-polygon-layer';
 import PathLayer from '../path-layer/path-layer';
 import * as Polygon from '../solid-polygon-layer/polygon';
 import {replaceInRange} from '../utils';
+import {assert} from '@deck.gl/core';
+import {PropTypeDef} from 'modules/core/src/lifecycle/prop-types';
+
+// TODO: not sure where this type belongs to
+export type MaterialProps = {
+  ambient: number;
+  diffuse: number;
+  shininess: number;
+  specularColor: [r: number, g: number, b: number];
+};
+
+/**
+ * Properties for `PolygonLayer`.
+ */
+export type PolygonLayerProps<DataT = any> = CompositeLayerProps<DataT> & {
+  /**
+   * Whether to draw an outline around the polygon (solid fill).
+   *
+   * Note that both the outer polygon as well the outlines of any holes will be drawn.
+   */
+  stroked: boolean;
+
+  /**
+   * Whether to draw a filled polygon (solid fill).
+   *
+   * Note that only the area between the outer polygon and any holes will be filled.
+   */
+  filled: boolean;
+
+  /**
+   * Whether to extrude the polygons
+   *
+   * Based on the elevations provided by the `getElevation` accessor.
+   *
+   * If set to `false`, all polygons will be flat, this generates less geometry and is faster
+   * than simply returning 0 from getElevation.
+   */
+  extruded: boolean;
+
+  /**
+   * Elevation multiplier.
+   *
+   * The final elevation is calculated by `elevationScale * getElevation(d)`.
+   * `elevationScale` is a handy property to scale all elevation without updating the data.
+   */
+  elevationScale: boolean;
+
+  /**
+   * Whether to generate a line wireframe of the hexagon.
+   *
+   * The outline will have "horizontal" lines closing the top and bottom polygons and a vertical
+   * line (a "strut") for each vertex on the polygon
+   */
+  wireframe: boolean;
+
+  /**
+   * The units of the line width, one of `meters`, `common`, and `pixels`.
+   *
+   * @see Unit.
+   */
+  lineWidthUnits: Unit;
+
+  /**
+   * The line width multiplier that multiplied to all outlines of `Polygon` and `MultiPolygon`
+   * features if the stroked attribute is true.
+   */
+  lineWidthScale: number;
+
+  /**
+   * The minimum line width in pixels.
+   *
+   * @default 0
+   */
+  lineWidthMinPixels: number;
+
+  /**
+   * The maximum line width in pixels
+   *
+   * @default Number.MAX_SAFE_INTEGER
+   */
+  lineWidthMaxPixels: number;
+
+  /**
+   * Type of joint. If `true`, draw round joints. Otherwise draw miter joints.
+   *
+   * @default false
+   */
+  lineJointRounded: boolean;
+
+  /**
+   * The maximum extent of a joint in ratio to the stroke width.
+   *
+   * Only works if `lineJointRounded` is false.
+   *
+   * @default 4
+   */
+  lineMiterLimit: number;
+
+  lineDashJustified: boolean;
+
+  /** Called on each object in the data stream to retrieve its corresponding polygon. */
+  getPolygon: AccessorFunction<DataT, any>;
+
+  /**
+   * Fill collor value or accessor.
+   *
+   * @default [0, 0, 0, 255]
+   */
+  getFillColor?: Accessor<DataT, Color>;
+
+  /**
+   * Line color value or accessor.
+   *
+   * @default [0, 0, 0, 255]
+   */
+  getLineColor?: Accessor<DataT, Color>;
+
+  /**
+   * Line width value or accessor.
+   *
+   * @default [0, 0, 0, 255]
+   */
+  getLineWidth?: Accessor<DataT, number>;
+
+  /**
+   * Elevation valur or accessor.
+   *
+   * Only used if `extruded: true`.
+   *
+   * @default 1000
+   */
+  getElevation: Accessor<DataT, number>;
+
+  /**
+   * This property has been moved to `PathStyleExtension`.
+   *
+   * @deprecated
+   */
+  getLineDashArray: Accessor<DataT, number> | undefined;
+
+  /**
+   * If `false`, will skip normalizing the coordinates returned by `getPolygon`.
+   *
+   * **Note**: This prop is experimental
+   *
+   * @default true
+   */
+  _normalize: boolean;
+
+  /**
+   * Specifies the winding order of rings in the polygon data.
+   *
+   * **Note**: This prop is experimental
+   *
+   * @default 'CW'
+   */
+  _windingOrder: 'CW' | 'CCW';
+
+  /**
+   * Material props for lighting effect.
+   *
+   * @see https://deck.gl/docs/developer-guide/using-lighting#constructing-a-material-instance
+   */
+  material?: true | MaterialProps;
+};
 
 const defaultLineColor = [0, 0, 0, 255];
 const defaultFillColor = [0, 0, 0, 255];
 
-const defaultProps = {
+const defaultProps: Record<string, PropTypeDef> = {
   stroked: true,
   filled: true,
   extruded: false,
@@ -57,7 +231,13 @@ const defaultProps = {
   material: true
 };
 
-export default class PolygonLayer extends CompositeLayer {
+export default class PolygonLayer<
+  DataT = any,
+  PropsT extends PolygonLayerProps<DataT> = PolygonLayerProps<DataT>
+> extends CompositeLayer<PropsT> {
+  static layerName = 'PolygonLayer';
+  static defaultProps = defaultProps;
+
   initializeState() {
     this.state = {
       paths: []
@@ -69,6 +249,8 @@ export default class PolygonLayer extends CompositeLayer {
   }
 
   updateState({oldProps, props, changeFlags}) {
+    assert(this.state);
+
     const geometryChanged =
       changeFlags.dataChanged ||
       (changeFlags.updateTriggersChanged &&
@@ -93,9 +275,9 @@ export default class PolygonLayer extends CompositeLayer {
     }
   }
 
-  _getPaths(dataRange = {}) {
+  _getPaths(dataRange: {startRow?: number; endRow?: number} = {}) {
     const {data, getPolygon, positionFormat, _normalize} = this.props;
-    const paths = [];
+    const paths: {path: number[]}[] = [];
     const positionSize = positionFormat === 'XY' ? 2 : 3;
     const {startRow, endRow} = dataRange;
 
@@ -163,6 +345,8 @@ export default class PolygonLayer extends CompositeLayer {
       updateTriggers,
       material
     } = this.props;
+
+    assert(updateTriggers && this.state);
 
     const {paths, pathsDiff} = this.state;
 
@@ -263,6 +447,3 @@ export default class PolygonLayer extends CompositeLayer {
   }
   /* eslint-enable complexity */
 }
-
-PolygonLayer.layerName = 'PolygonLayer';
-PolygonLayer.defaultProps = defaultProps;
